refactor(mountains): extract terrain samplers + heightMap into mountains/terrain.ts
Move the analytic terrain height field — getTerrainHeight / getTerrainGradient /
getDownhillDirection, the shared heightMap cache, debugHeightMap, and the
TerrainVec2 type — into src/mountains/terrain.ts (pure math, no THREE/DOM). This is
the physics seam: getTerrainHeight stays byte-identical to the mesh-vertex formula,
and heightMap remains a single ES-module singleton that terrain-mesh.ts will
pre-populate.

The hub imports and re-exports them; TerrainVec2 stays on the public ./mountains.js
surface. Verified byte-identical: full Node suite green, physics-invariant harness
OK (coasting baseline unchanged), verify 49/49, tsc + lint clean.
